Move and rename files getting rid of automl
